Add a hover state to each row
This is potentially buggy. I had to do some unique things (like add a hover state to both the room-name and reserved-event divs) to allow for the title to still show (the popup text on hover) when hovering over a reserved-event. There’s got to be a better/cleaner way. But it’s good enough for MVP.